ai-server/requirements.txt
# FastAPI and server dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Google Gemini AI
google-generativeai==0.3.2

# Audio processing
librosa==0.10.1
soundfile==0.12.1
# webrtcvad==2.0.10  # Optional - requires C++ compiler
pydub==0.25.1

# Computer vision and face analysis
opencv-python==4.8.1.78
mediapipe==0.10.8
# deepface==0.0.79  # Optional - requires dlib which needs C++ compiler
# dlib==19.24.2     # Optional - requires C++ compiler and cmake

# NLP and text processing
spacy==3.7.2
nltk==3.8.1
textblob==0.17.1
transformers==4.35.2
torch==2.1.1

# Speech recognition
SpeechRecognition==3.10.0
# whisper==1.1.10  # Commented out due to compatibility issues on Windows

# Data processing
numpy==1.24.3
pandas==2.0.3
scikit-learn==1.3.2

# HTTP client
httpx==0.25.2
requests==2.31.0

# Environment and configuration
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# Logging and monitoring
loguru==0.7.2

# File processing
Pillow==10.1.0
python-magic==0.4.27
PyPDF2==3.0.1
python-docx==0.8.11

# Database (if needed)
motor==3.3.2
pymongo==4.6.0

# Caching
redis==5.0.1

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1